+++ callback/handler/donutcreate/handler_test.go
@@ -21,7 +21,7 @@ func TestCreate(t *testing.T) {
 
 	cases := []struct {
 		userId int64
-		amount int
+		amount float64
 
 		object json.RawMessage
